Transpose input data requirements
We no longer keep a table with many columns for the various input
formats. Instead, there's one big table to enter measurements. This
table has exactly three columns: a date, a key, and a value.

Entering data in this fashion has multiple benefits: we don't have to
enter useless NA values for non-existing measurements, it's not as
error-prone to enter measurements due to the large number of columns,
and we now can keep a central file with all relevant data.